@luotao1 2019-10-22
Please cherry-pick to release/1.6