@wzzju 2019-10-22
> Some minor conv int8 code reorders in order to get closer to 1.5.0 tag performance.
> **Test machine:** Xeon Gold 6248
> **Dataset:** Full validation dataset (50k images)
> **Target release:** ?
> 
> ```
> MOBILENET V1
> Develop:
> I1019 08:51:21.215854 215101 tester_helper.h:490] FP32: avg fps: 76.7701, avg latency: 13.0259 ms
> I1019 08:51:21.215874 215101 tester_helper.h:490] INT8: avg fps: 218.5339, avg latency: 4.5759 ms
> 
> TAG 1.5.0:
> I1019 09:13:37.712121 215210 tester_helper.h:469] FP32: avg fps: 74.9883, avg latency: 13.3354 ms
> I1019 09:13:37.712128 215210 tester_helper.h:472] INT8: avg fps: 230.3150, avg latency: 4.3419 ms
> 
> Fix:
> I1019 09:42:01.112879 230507 tester_helper.h:490] FP32: avg fps: 76.5687, avg latency: 13.0602 ms
> I1019 09:42:01.112900 230507 tester_helper.h:490] INT8: avg fps: 225.9895, avg latency: 4.4250 ms
> ```
> 
> ```
> MOBILENET V2
> Develop:
> I1019 14:20:57.676785 361283 tester_helper.h:490] FP32: avg fps: 89.4302, avg latency: 11.1819 ms
> I1019 14:20:57.676810 361283 tester_helper.h:490] INT8: avg fps: 198.5545, avg latency: 5.0364 ms
> 
> TAG 1.5.0:
> I1019 14:36:31.470321 361845 tester_helper.h:469] FP32: avg fps: 89.9544, avg latency: 11.1167 ms
> I1019 14:36:31.470330 361845 tester_helper.h:472] INT8: avg fps: 210.2631, avg latency: 4.7559 ms
> 
> Fix:
> I1019 15:27:58.629585 371148 tester_helper.h:490] FP32: avg fps: 88.2552, avg latency: 11.3308 ms
> I1019 15:27:58.629611 371148 tester_helper.h:490] INT8: avg fps: 204.0604, avg latency: 4.9005 ms
> ```

Why does the performance of fp32 become more bad?