@grygielski 2019-10-22
@wzzju It's just a measurement noise, fp32 performance remained the same.